+++ Covenant/Models/Grunts/Folder.cs
@@ -2,34 +2,10 @@
 using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
-using System.Diagnostics.CodeAnalysis;
 using Newtonsoft.Json;
 
 namespace Covenant.Models.Grunts
 {
-    public interface IFolderFileNode
-    {
-        [Key, DatabaseGenerated(DatabaseGeneratedOption.Identity)]
-        public int Id { get; set; }
-
-        [Required]
-        public string FullName { get; set; }
-        [Required]
-        public string Name { get; set; }
-        public long Length { get; set; }
-        public DateTime CreationTime { get; set; }
-        public DateTime LastAccessTime { get; set; }
-        public DateTime LastWriteTime { get; set; }
-
-        public int? ParentId { get; set; }
-        [JsonIgnore, System.Text.Json.Serialization.JsonIgnore]
-        public Folder Parent { get; set; }
-
-        public int GruntId { get; set; }
-        [JsonIgnore, System.Text.Json.Serialization.JsonIgnore]
-        public Grunt Grunt { get; set; }
-    }
-
     public abstract class FolderFileNode
     {
         [Key, DatabaseGenerated(DatabaseGeneratedOption.Identity)]
@@ -37,7 +13,6 @@ namespace Covenant.Models.Grunts
 
         [Required]
         public string FullName { get; set; }
-        [Required]
         public string Name { get; set; }
         public long Length { get; set; }
         public DateTime CreationTime { get; set; }